change order of locator class usage
